refactor: separate functions into files
