+++ domain/random.go
@@ -1,0 +1,8 @@
+package domain
+
+import "math/rand"
+
+// Random returns a random number
+func Random() int {
+	return rand.Intn(100000000)
+}
