+++ domain/replace_cpf.go
@@ -1,0 +1,10 @@
+package domain
+
+import "strings"
+
+// CpfReplace returns the CPF received in a single format
+func CpfReplace(cpf string) string {
+	cpfReplace := strings.Replace(cpf, ".", "", 2)
+	cpfReplace = strings.Replace(cpfReplace, "-", "", 1)
+	return cpfReplace
+}
